分布式事务学习

+++ distributed-transaction/dtx-seata-saga/dtx-seata-saga-account/src/main/java/com/deng/seata/saga/account/facade/request/BalanceInfo.java
@@ -12,5 +12,5 @@ import java.io.Serializable;
 @Data
 public class BalanceInfo implements Serializable {
     private String name;
-    private int amount;
+    private Long amount;
 }
